Move layout algorithms to own module
This change encapsulates the layout algorithms in their own module.
The first benefit is that the long layout.cpp becomes shorter. The other
benefit is that it will be easier to implement new layout algorithms
(which is requested quite frequently).
